+++ src/main/java/com/library/domain/BookCopies.java
@@ -12,15 +12,15 @@ import javax.persistence.*;
 public class BookCopies {
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO)
-    private long bookId;
-    private long titleId;
+    private int bookId;
+    private int titleId;
     @Column(name = "BOOK_STATUS")
     private String bookStatus;
 
     public BookCopies() {
     }
 
-    public BookCopies(long titleId, String bookStatus) {
+    public BookCopies(int titleId, String bookStatus) {
         this.titleId = titleId;
         this.bookStatus = bookStatus;
     }
@@ -30,11 +30,11 @@ public class BookCopies {
     }
     @ManyToOne
     @JoinColumn(name = "TITLE_ID")
-    public long getTitleId() {
+    public int getTitleId() {
         return titleId;
     }
 
-    public long getBookId() {
+    public int getBookId() {
         return bookId;
     }
 }
